Skip submodules when listing files
If we call git annex info on a submodule path (which is reported by
ls-tree), git-annex will oblige, but the result will not be a file
information. Previously, that would cause a crash when trying to
report file size (on a thing that's not a file and has no such
property).