Use datalad status to create a dataset file listing
In a glaring omission from my side previously, datalad status can be
used directly on a dataset to report on file paths and sizes, annexed
or not. This both simplifies the code and brings a massive speed
improvement, bringing the time down to < 10 s for datasets with
ca. 40k files.

Prompted by a comment in PR #48 -- at least I can treat that one as an
exercise in git-annex usage...

Suggested-by: Michael Hanke <michael.hanke@gmail.com>