Include the ability to list files in extract_selected
With the newly (for me personally) found speed and capabilities in
datalad status, there is no reason not to include generating a file
list in extract_selected. It is, however, kept as a CLI option.

The list_files script is converted to an importable module providing a
list_files function (that uses status to create a list of file info
and returns a generator that yields catalog-ready items), but keeps is
CLI (thanks to `if __name__ == "__main__"`).