Add a superdataset configuration (subdataset tab default)
Add a dataset-level configuration file to be used for sfb- and
project-superdatasets. It is a copy of the catalog-level configuration
file with one addition, dataset_options.default_tab = subdatasets.

This allows us to work with a default that shows a dataset with
content tab open (which is a preferred way to show project
subdatasets) while making an "exception" for the two top levels (sfb
and project) which would show subdatasets first.