Update superdataset extraction procedure
CFF extractor is run on the superdataset, in addition to the studyminimeta
extractor. Current config is sufficient to seamlessly merge the information
from the two extractors. Non-overlapping fields that matter are: funding
(only studyminimeta), doi (only CFF).

A hardcoded license name for the superdataset is added. This allows us to
have a custom text displayed in the license field, without relying on the
metadata format limitations (studyminimeta has no license property, CFF
understood strictly should only contain SPDX identifiers).

As a side note, since we are already making hardcoded additions, we could
ditch studyminimeta in favour of CFF and use additions for funding (just
like we do for project datasets), but I am stopping short of that now.

Closes #80