Fix backbutton display bug after switching to history mode
With history mode and the href properties added to anchor tags,
and also using preventDefault with the onclick event of achor tags,
a bug was introduced where the dataset backbutton didn't display for
a dataset when it should, i.e. for a subdataset navigated to from the
home page. When using the browser back button, the dataset backbutton
just flashed quickly. The fix was this was shifting the 'show_backbutton'
property to a computed property (showBackButtonComp)