Fix an incorrectly formatted doi (in metadata)
This is a manual edit of a publication DOI in Z03 motor-assessment-derived
dataset metadata, adding the missing https://doi.org/ to make the link
point to the publication.

This is due to an oversight in tabby-utils, which has now been fixed
to prevent reintroducing the error. No other datasets were affected.

See:
https://github.com/sfb1451/tabby-utils/commit/949cb18d6d46a248f885e01a270f44a64411705c