Allow manual additions in extract_selected, add notebooks
Lookup of `manually_entered` toml file (similar to the one performed
by other scripts) is added to `extract_selected`, used mainly for
running select extractors on a project subdataset.

The toml file gains an "additional" table, with datasets IDs used as
keys. The idea is that it can be used to add already formatted
metadata for specific datasets, by their ID. One entry is added, a
"notebooks" array for the B03 externally cued movements (its value was
previously edited in by hand, directly to catalog metadata).

Closes #90 (although it does not address the hypothetical case of
tabby addition, which will need to be addressed separately).