Remove handling of studyminimeta
We used to extract studyminimeta metadata from the superdataset, and
had a postprocessing function that applied a convention to the funding
property. These are no longer needed (superdataset funding is
hardcoded). I am also not aware of any dataset which uses the
studyminimeta format, so all specific code can be retired.