Add ad-hoc code for initial dataset tagging
Datasets need to have project names assigned as keywords. These
typically need to come from "outside". Here, we add a toml file with a
dataset - project mapping. Helper scripts used to generate (part of)
the file are included for future reference.

Later we could e.g. add a flag to the dataset extraction script which
can be used to provide the mapping when extracting for the first time.