Add ability to look up GIN doi
This feature becomes part of extract_selected. If requested (via
--gindoi), and if the dataset's origin is on GIN, the code gets the
dataset landing page and parses it using BeautifulSoup to obtain a GIN
DOI.

This can be useful if a dataset has a DOI from GIN but it is not
declared in any of the sources (for example, a CFF file is required to
obtain a DOI from GIN, but people rarely add the DOI there once
obtained).

Closes #97