Remove incorrect (?) lines carried over from upstream
This fixes what seems to be a bug introduced from catalog upstream via
4aec3ec4. The removed lines are apparently left over from editing, as
they are hanging there. When they are left in, the lack of optional chaining
(`?.`, which is used above) causes errors whenever the additional
display does not include a homepage.

The behaviour was weird: dataset title carried over when opening a
subdataset, and the subdataset could not be opened in a new tab.