Display subdataset titles instead of their directory names
The subdataset card is modified so that the subdataset title is shown
as the card header (the big text linking to the subdataset page),
replacing directory name. Sorting and filtering are adjusted
accordingly. Sorting is made case-insensitive.

The directory name is used as the card header if the title is not
available, and always shown in small text as "directory".

To achieve this, subdatasets_ready() function is modified to add a
display_name property to each subdataset object (similar to what
dataset_ready does to the superdataset). The property is needed twice:
for sorting (js code) and for display (html template), so it seemed
appropriate to add the property instead of having to use the ternary
operator twice.

Subdataset sorting uses localeCompare() on the names, which seems to
be the recommended method to achieve case-insensitive comparison:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/localeCompare

The fact that we only define the name for available datasets does not
seem to be a problem, as the sortedSubdatasets() is called only after
filteredSubdatasets() which filters by the available property.

Closes #96