Add DOI badge to subdataset view
A DOI badge will be shown in the subdataset card if a DOI is
available. Uses https://badgen.net/

Badgen was chosen over shields.io mostly because it is slightly easier
to format the URL (using encodeURICOmponent instead of replacing
hyphens and underscores). It also promises to be faster and leaner.

A future rework could implement a simplified shield representation
using divs with backgrounds (similar to what GIN does) or include the
badgen JS library (permissive license) to remove reliance on external
service.

Closes #22