Add ability to override binder URL - a workaround
This is a much simplified version of 72fa4adb and achieves
similar (but not the same) goal via an off-the-schema property.

If dataset metadata defines `notebooks[0].binder_link`, it will be
used, unchanged, for the "open with binder" button. The schema is not
changed, so other notebook properties need to be defined (but they can
be empty strings).

This is a temporary solution, and should be replaced with a proper
implementation - schema change, and possibly refactoring the link
generation. Ideally, the solution would be implemented upstream. For
now, this must be sufficient.