Add hardcoded binder link for delayed_center-out_uncertainty
This adds a binder link for one of the A06 datasets. It is a fully
custom link - dedicated repo with Docker environment and a demo
notebook. It uses an undocumented (off-schema) property `binder_link`
as an override, thanks to local catalog customisation (hopefully
temporary).

This particular binder demo is a (limited) showcase of re-executing
simulation code to generate a published figure.